added Vault support and automatic timed promotions, fixed some permission checks and added /playtime command

+++ CommandsEX/src/com/github/zathrus_writer/commandsex/commands/Command_cex_hyes.java
@@ -11,6 +11,7 @@ import com.github.zathrus_writer.commandsex.CommandsEX;
 import com.github.zathrus_writer.commandsex.SQLManager;
 import com.github.zathrus_writer.commandsex.helpers.Home;
 import com.github.zathrus_writer.commandsex.helpers.LogHelper;
+import com.github.zathrus_writer.commandsex.helpers.Permissions;
 import com.github.zathrus_writer.commandsex.helpers.Utils;
 
 public class Command_cex_hyes extends Home {
@@ -32,22 +33,24 @@ public class Command_cex_hyes extends Home {
 				return true;
 			}
 			
-			try {
-				// remove homes
-				String[] qMarks = new String[clearHomesConfirmation.get(pName).size()];
-				Arrays.fill(qMarks, "?");
-				SQLManager.query("DELETE FROM " + SQLManager.prefix + "homes WHERE id_home IN ("+ Utils.implode(qMarks, ",") +")", clearHomesConfirmation.get(pName));
-				
-				// inform player
-				LogHelper.showInfo("homeClearDone", sender);
-				
-				// clear the IDs cache
-				clearHomesConfirmation.remove(pName);
-			} catch (Throwable e) {
-				// something went wrong with our SQL...
-				LogHelper.showWarning("internalError", sender);
-				LogHelper.logSevere("[CommandsEX] " + _("dbReadError", ""));
-				LogHelper.logDebug("Message: " + e.getMessage() + ", cause: " + e.getCause());
+			if (Permissions.checkPerms(player, "cex.hyes")) {
+				try {
+					// remove homes
+					String[] qMarks = new String[clearHomesConfirmation.get(pName).size()];
+					Arrays.fill(qMarks, "?");
+					SQLManager.query("DELETE FROM " + SQLManager.prefix + "homes WHERE id_home IN ("+ Utils.implode(qMarks, ",") +")", clearHomesConfirmation.get(pName));
+					
+					// inform player
+					LogHelper.showInfo("homeClearDone", sender);
+					
+					// clear the IDs cache
+					clearHomesConfirmation.remove(pName);
+				} catch (Throwable e) {
+					// something went wrong with our SQL...
+					LogHelper.showWarning("internalError", sender);
+					LogHelper.logSevere("[CommandsEX] " + _("dbReadError", ""));
+					LogHelper.logDebug("Message: " + e.getMessage() + ", cause: " + e.getCause());
+				}
 			}
 		} else {
 			// no database, no homes
